// app/Models/Users/users.ts
import { usersSchool } from './usersSchool';

export class users{

    id: number;
    employeeId: number;
    username: string;
    password: string;
    email: string;
    locale: string;
    schoolId: number;
    arSchoolName: string;
    enSchoolName: string;
    yearId: number;
    yearName: string;
    usersSchool: usersSchool;
    roleName: string;
    isSuperAdmin:string;
    isActive:string;
    title:string;
    titleAr:string;
    active:number;
    name:string;
}